refactor(phase-5): extract navigateToStep + handleBranchTarget into tour-engine
Move the two largest navigation orchestrators out of tour-provider.tsx into
pure module-level functions under packages/core/src/lib/tour-engine/. The
extracted impls receive a TourEngineContext (refs + getters) so async
navigation reads fresh state across await boundaries.

- new types/tour-reducer.ts holds TourAction + TourReducerState so engine
  modules can dispatch without importing provider implementation
- new lib/tour-engine/context.ts defines TourEngineContext (getters/refs)
- new lib/tour-engine/helpers.ts hoists the pure step-visibility utilities
  (buildCallbackContext, evaluateStepWhen, findNextVisibleStepIndex,
  findNearestVisibleStepIndex, isNavigationNeeded)
- new lib/tour-engine/navigate-to-step.ts owns navigateToStepImpl and its
  advancePastHiddenStep helper
- new lib/tour-engine/handle-branch-target.ts owns handleBranchTargetImpl
- tour-provider.tsx keeps reducer, flow restore, and prev; navigateToStep
  and handleBranchTarget are now thin useCallback wrappers that dispatch
  through engineContextRef.current

Behavior preserved: all 928 existing core tests pass. Provider drops from
1803 LOC; navigateToStep + handleBranchTarget complexity ignores move
from provider to engine impls.
